Keep test/ of the load-path
Bundling third-party libraries is problematic because the bundled
copies tend not to get updated.  A user who has both the upstream
package and some other package that bundles it installed, might
then end up loading the bundled and outdated version.

The file ".nosearch" instructs some tools to keep the containing
directory of the `load-path'.  However the best approach would be
to not bundle any third-party libraries at all.