Limit directory-local values to emacs-lisp
In order to create this and the preceding commit I had to tell Emacs
that I didn't want to set the value of `elisp-lint-indent-specs` about
a dozen times.  If the scope of this setting had been limited to Elisp,
then I would only have had to do it once or twice.